+++ CHANGELOG.md
@@ -11,6 +11,9 @@
 -   Changed `shout()` and `whisper()` to cost 1 energy point.
     -   This helps prevent infinite loops.
     -   The energy point is only deducted if a bot has a listener for the event.
+-   Changed `experiment.localPositionTween()` and `experiment.localRotationTween()` to take an options object as the 4th parameter instead of the easing options.
+    -   This options object is able to accept easing and duration values.
+    -   See the docs for examples.
 
 ### :bug: Bug Fixes